Add a function to print whole types.
With this, it's possible to get more information than the simplified
types shown at the REPL. The types are as unwieldy as ever, but you
have to ask for them explicitly.

This is not an adequate long-term solution, and will be replaced by a
"Type Explorer" tool.

original commit: 70e1d63bed2ac85313d048cc9c359ae43330a7e3