Fix type error reporting for case-> with * domains
In the case that a case-> type included a case with a *
domain and had additional branches, a use of `apply` with
that type would fail to report a type error. This was
because the type of the applied list was ignored for type
error generation.

Closes PR 13893

original commit: 328956e8b5cd9b9c6256b1eac4f46aec98ce2a5d