Add type annotations for GUI libraries
The types for racket/gui are substantially complete, but
there are still some missing classes and methods.

Some interesting future work would be to ensure type coverage
by comparing the class members vs. the types in the environment.
That could be run as a test to ensure it stays up to date.

Most types for the framework library are still missing.

original commit: 00a3b7cb149ae35c63b9eb64df34d87424407a10