Temporarily disable TR class unit test
This unit test is broken due to an issue with using
the : form for external names in classes. Since : checks
if the id is bound, it doesn't work if there is an internal
name different from the external name (since only the
internal one is in scope in the class body).

Haven't decided whether to put annotations on internal
identifiers instead or to not run the check for class
member names.

original commit: fe811e33aaf50fe68ad2ff14a7c46685a03e1f19