Catch syntax lifting at the top-level in TR
This is helpful for typechecking forms that lift
expressions such as object instantiation or contracted
values when they are used at the top-level.

It's not normally an issue for typechecking modules
because lifting is delimited by the module extent so that
normal local-expansion will cover it.

original commit: 00470e3e1a09f88a32c6b20682b577669655a2c4