Don't track syntax with unknown source locations
Avoids recording original syntaxes with a source or position
of #f, since these are usually intermediate syntax pairs that
don't correspond to actual source syntax.

For example, in a (#%module-begin . rst) in a submodule, the
underlying pair may contain an identifier and a cdr that is a
syntax object. That cdr will have #f source and position and also
does not correspond to any source syntax.

Leaving such syntax objects in the table means there are spurious
hits later when a syntax object with an unknown source location
is looked up.

Closes PR 14389

original commit: 06bd5d36af88a99d2f76b65c073da21d63b5ee1f