repairs for top-level interaction
After local expansion, switch to an "outside" view of the expanded
term by using `syntax-local-introduce`, then switch back just before
producing the result for `#%module-begin` or `#%top-interaction`.

This switch is only a minor simplication for modules. It matters more
for top-level interactions, where each expansion of
`#%top-interaction` is a different macro invocation (so the "outside"
view is the consistent one across interactions).