Fix typed provides used in some kinds of modules
When identifiers provided by typed modules were used in
certain submodules of the form (module* n #f ...) or were
used by modules implemented in a language defined by TR,
the wrong redirection was used in the expansion.

The reason was because TR's identifier redirection decided
whether it was in a typed or untyped context at module visit
time, but that's too early in the cases above.

(because TR's #%module-begin may not have begun expanding yet)

Use the new prop:syntax-local-value feature to construct a
redirection that acts like a rename-transformer but which delays
the decision to use the typed or untyped identifier until
expansion time.

Closes GH issue #163 and #181

Closes PR 15118