change source location for optimizer test
For some reason, the optimizer is logging an entire case-lambda
expression when it should be logging the first branch.

This commit does not fix the problem, but changes the expected log
message in a test to avoid a test failure. (The optimizer does seem to
be optimizing correctly --- its just reporting the wrong thing.)

The bad source location seems to be coming from
`typed-racket/optimizer/dead-code.rkt`:

```
          #,@(for/list ((formals (in-syntax #'(formals ...)))
                        (bodies  (in-syntax #'(bodies ...)))
                        #:when (dead-lambda-branch? formals))
                (log-optimization
                  "dead case-lambda branch"
                  "Unreachable case-lambda branch elimination."
                  formals)
                #`(λ #,formals . #,bodies))))))
```

Things I've tried:
- checking the value of `formals` inside the loop; it is correctly the
  different arg-specs for the function
- checking `log-optimization`; that function calls `locate-stx` on its
  third argument; after adding prints it looks like `locate-stx` is not
  returning the bad syntax object that appears in the log message
- renaming `formals` to something new; no effect