static-contract: improve or/sc optimization
Allow more optimizations to the branches in an `or/sc`.

Before this PR, the function `make-trusted-side-reduce` would never
optimize the branches of an `or/sc`. That was overly conservative
because some of the optimizations do not change the contract constructor,
for example:

- it's safe to change `(or/sc integer? (-> boolean? boolean?))`
  into `(or/sc integer? (-> boolean? any))`
  if the contract is guarding typed values flowing to untyped code

After this PR, `make-trusted-side-reduce` gets a new argument to help it
decide what optimizations are safe to apply.

- in contrast to the last example, it is NOT safe to change
  `(or/sc integer? (-> boolean? boolean?))`
  into `(or/sc any/c (-> boolean? boolean?))`
  if the contract is guarding typed values flowing to untyped code.
  The reason is, that contract would further optimize to `any/c` and
   then any typed function flowing through the contract would not be
   protected against non-boolean arguments

The general rule is to apply any optimization that doesn't change the
contract constructor of the branches. More low level:
if `(or/sc sc0 sc1)` optimizes to `(or/sc sc0+ sc1)` then it must be true
that `(contract-first-order [[sc0]]) = (contract-first-order [[sc0+]])`.

(Where [[sc0]] is the instantiation of the static contract)