type-contract: flatten contract for unions
Typed Racket's internal representation of unions represents a type
`(U T ...)` as something like two unions, `(U (U Ti ...) Tj ...)`,
where the `Ti` are the "base types" and the `Tj` are the other types.

This commit does two things about this:

1) changes the type->contract function to flatten the contract
 for union types.

- before: `(or/sc (or/sc Ti ...) Tj ...)`,
- after : `(or/sc Ti ... Tj ...)`

2) adds a case to the static-contracts optimizer to flatten a nested
`or/sc` if the contracts inside are either all-flat or all-chaperones.