attempt to fix types for `raise-argument-error`
currently this fix does not work in that attempts to use
`raise-argument-error` in TR throw exceptions, so here you go.

Based on various examples, I extended `->key` in `base-abbrev.rkt` to
support rst pattern variable since the new `raise-argument-error`
defintion has a rest argument.

I then move the old `raise-argument-error` type to `base-env.rkt` and
replace the type abstraction with the extended `->key` abstraction.

My interpretation of the `->key` pattern is that since `#:more-info` is
an optional keyword it needs to have the following signature
`#:more-info (Un -String (-val #f)) #t` since it has default argument of
`#f` if no strings are provided. I think `#t` is used to indicate that
`#:more-info` is optional but I could be entirely wrong.