Preserve plambda property when typechecking expanded opt-lambdas
lambdas with optional arguments expand to a
  (let-values (((core) (lambda))) (case-lambda))
form, which is special-cased in tc-expr-unit. Properties added
to the expansion of typed-lambda (including type var information)
are attached to the whole let-values expression, rather than the
core lambda. To correctly parse type variables in the core lambda,
this patch re-adds the plambda property when passing it to typecheck.