Separate types of find-or-create-brush depending on first argument.
According to the racket/draw documentation [0], find-or-create-brush returns a
brush% or #f if the first argument is a string, but always returns a brush% if
the first argument is a color% instead.

The linked docs aren't very clear; according to the above link,
find-or-create-brush may also always return a brush% even if its first argument
is a string.

As a side note, the implementation of find-or-create-brush [1] seems to return a
black brush if the first argument is a string that isn't recognized as a color%.
Should the documentation be updated to reflect this or is this intentionally
left ambiguous?

[0]: https://docs.racket-lang.org/draw/brush-list_.html#%28meth._%28%28%28lib._racket%2Fdraw..rkt%29._brush-list~25%29._find-or-create-brush%29%29
[1]: https://github.com/racket/draw/blob/6ada7d8abdc664b4b4b42c6e2a7a3703d3794f9e/draw-lib/racket/draw/private/brush.rkt#L222