remove type anns for parameters in λs generated from kw/opt plambdas
an optional or keyword lambda is expanded to more than two lambdas where
syntax objects for the original paramaters are reused. This is problematic when
those parameters are typed with bound type variables in polymorphic functions,
because in generated lambdas those variables would become unbound.

This patch fixes the problem by throwing the types away before typechecking
the generated lambdas.

closes #940