repair & improve names generated from structs' definitions (#1067)
Previously when a struct's type name was different from the struct name, TR
did not attach its syntax transformer binding with the corresponding struct
info for the enclosing module, but did so for its exported transformer
binding. This was the root cause of the problems mentioned of #930.

With this change, the behavior is consistent, the type name of a structure can
be used as an alternative to its struct id. E.g. we can use a structure type name
in struct-copy, or use it as the super id in other structs' defintions.

Under the hood, this change includes a fair amount of refactoring:

1. Move existing helper functions and structures for struct-info from the
"jack-of-all-trades" utils/utils. Add more functions and structures. In particular, we
nows use three struct-info structures to handle three different situations of
generating bindings of a struct's type name and struct name.

2. Where the binding of a structure's type name for the enclosing module is
generated has moved in favor of the consistency.

3.the heavy-lifting job of provide handling has moved to def-binding.rkt, where
every binding structure takes care of its corresponding quad generation.

lastly, closes #930