bugfix for struct-related names cont'd:
Using the constructor name from the struct info to decide whether a struct name
can be used as its constructor is problematic, because in that case the
constructor name is the extra constructor name when the latter is
supplied. Instead, we should pass the constructor name from (define-) struct
forms directly.

This is also a spinoff bugfix from #1078.

These changes include small fixes for `define-struct/exec`

closes #1088