fix nondetermistic behavior of path handling in env+
1. previously, the order of the list of path elements to an object was
nondeterministic, due to relying on implementation of a hashmap. This change sorts
the list and ensure objects referencing the identifier with fewer path elements
always come out ealier. This enables the typechecker to use the list of paths in
reasonable and correct order.

e.g:
(struct foo ([a : Any]))
(define var : Any (foo 10))
(if (and (foo? var) (number? (foo-a var)))
    (number-a var)
    42)

when checking the then branch above, the typechecker now uses var@foo
then (foo-a var)@number to refine var. So the program now have the same
typechecking result as the following:

(struct foo ([a : Any]))
(define var : Any (foo 10))
(if (foo? var)
    (if (number? (foo-a var))
        (number-a var)
        42
    42)

closes #1098

2. fix two potential internal errors.