support prop:procedure
1. use intersection types for constructors of structures that are associated with prop:procedure
2. use define-typed-struct to implement define-struct/exec
3. handle substructuring of procedural structures properly:
  - a substructure of an procedural structure is also procedural
  - but such a substructure can specify a procedure of a function type
  potentially different from the parent one or a different field for its own
  prop:procedure

Implemention-wise, it is worth pointing out that how to pass the function type
for a base procedural structure to a substructure depends on where the base
structure is defined:
1. if it is defined in the same enclosing module as its substructures are, the
typechecker will make their contructors create executable instances, namely the
return types of the constructors will have intersection types (Inter Fun
SomeStructNameAsType) while checking property values. This process happens in
pass 2.

2. if it is defined in an imported module, the typechecker will grab the
procedure type from the Struct rep for the base structure and pass it to the
Struct rep for the substructue that is being checked, which happens before pass 1.