partially support structure properties in a structure definition at REPL
for expanded expressions of a structure definition, this repair ensures a
structure's properties are checked after the struct type is registered.

This change does not implement a complete static check for structure
properties. For example, if a prop:custom-write is given 10, then a dynamic
contract error will be triggered. This is because the make-struct-type will be
evaluated before registering struct type and checking proper values.