rewrite type-checking values for prop:procedure
Since a value for `prop:procedure` attached to a structure also changes the type
of its constructor, which can be used in other top-level bindings, we need to
gather type which the value for `prop:procedure` relates to before pass 1.

Also, support for lambdas with keyword paramenters were added in this commit.
Note that, similar to regular lambdas, type variables are not allowed to appear
in keyword parameters.