Use new Resyntax --output-to-file option
This should fix the review json artifact issue. The artifact currently uploaded by the analysis step is malformed because it includes debug output from Resyntax in addition to the intended json, see the `resyntax-review` artifact in [this workflow run](https://github.com/racket/typed-racket/actions/runs/2945825191) for an example of the issue. This commit should fix that, hopefully.