Make PLT_TR_CONTRACTS easier to toggle in CI
This changes PLT_TR_CONTRACTS so that the legal values are `"true"`, `"false"`, `""`, and unset, and only the first of those choices results in `enable-contracts?` being true. Previously, setting `PLT_TR_CONTRACTS` to any string at all (including explicitly setting it to the empty string) would enable contracts.

I made this change so I could simplify the CI setup a little, since this lets the `${{ matrix.enable-contracts }}` expression be passed directly in the environment.