type-contract: merge vector and hash types earlier (#1282)
Context:
  (Vector T) is implemented as the union of two types. Enforcing the
  type with two contracts is inefficient; we want one merged contract.
  Same for (HashTable K V), which is implemented as three types.

This commit replaces the old static-contract merging with a type
merging.

The problem with static contract merging is that it happens after
the "typed-side" optimization kicks in --- which meant that two
types with equal components could end up as static contracts with
unequal components. This was a huge practical problem in the GTP
benchmark `dungeon` ... the fix lowers the runtime for one loop
in the worst case down from 30 seconds to 1/2 second.