Propagate a trivial expected type for recursive lambda if neccessary.
This does not apply an expected type to the actual function body,
it just gives a type to the function when referred to recursively.

This addresses a problem with inference for let loop in lambda as
the argument to `thread` now that `thread` is a keyword function and
therefore lifts its arguments to a `let` ahead of the function.

Fixing the relevant problem also requires improving the type of `thread`.