Reduce per-PR Resyntax limits
Resyntax seems to occasionally take an extremely long time to analyze the repository (over four hours), and it seems to fail at the final step of creating a pull request in these cases. The error message mentions "bad credentials" so I suspect the API token given to the workflow is already expired by the time Resyntax is ready to create a PR.

This PR reduces the amount of work Resyntax can do per pull request in the hopes of making it do less work and finish quicker.